+++ src/Components/Welcome.js
@@ -9,8 +9,11 @@ export default function Welcome() {
   const navigate = useNavigate();
 
   function logOut() {
-    localStorage.clear();
-    navigate("/login")
+    const confirm = window.confirm('Are you sure to logout?');
+    if (confirm) {
+      localStorage.clear();
+      navigate("/login")
+    }
   }
 
   return (
